move createHttpLink stuff to HttpLink
